+++ clang/unittests/AST/ASTExprTest.cpp
@@ -108,3 +108,41 @@ TEST(ASTExpr, InitListIsConstantInitialized) {
   (void)FooInit->updateInit(Ctx, 2, Ref);
   EXPECT_FALSE(FooInit->isConstantInitializer(Ctx, false));
 }
+
+TEST(ASTExpr, NoUnsignedOverflowICE) {
+  std::unique_ptr<ASTUnit> AST = buildASTFromCode(R"cpp(
+    unsigned u = 1u * 8u;
+  )cpp");
+  ASTContext &Ctx = AST->getASTContext();
+  const VarDecl *Var = getVariableNode(AST.get(), "u");
+
+  const std::optional<llvm::APSInt> ICE =
+      Var->getInit()->getIntegerConstantExpr(Ctx);
+  EXPECT_TRUE(ICE.has_value());
+  EXPECT_TRUE(ICE->isUnsigned());
+  EXPECT_EQ(8u, *ICE);
+
+  const std::optional<llvm::APSInt> ICEOverflowCheck =
+      Var->getInit()->getIntegerConstantExpr(Ctx, nullptr, true);
+  EXPECT_TRUE(ICEOverflowCheck.has_value());
+  EXPECT_TRUE(ICEOverflowCheck->isUnsigned());
+  EXPECT_EQ(8u, *ICEOverflowCheck);
+}
+
+TEST(ASTExpr, CheckForUnsignedOverflowICE) {
+  std::unique_ptr<ASTUnit> AST = buildASTFromCode(R"cpp(
+    unsigned u = 1'000'000'000u * 8u;
+  )cpp");
+  ASTContext &Ctx = AST->getASTContext();
+  const VarDecl *Var = getVariableNode(AST.get(), "u");
+
+  const std::optional<llvm::APSInt> ICE =
+      Var->getInit()->getIntegerConstantExpr(Ctx);
+  EXPECT_TRUE(ICE.has_value());
+  EXPECT_TRUE(ICE->isUnsigned());
+  EXPECT_EQ(3'705'032'704u, *ICE);
+
+  const std::optional<llvm::APSInt> ICEOverflowCheck =
+      Var->getInit()->getIntegerConstantExpr(Ctx, nullptr, true);
+  EXPECT_FALSE(ICEOverflowCheck.has_value());
+}